requirements.txt
torch>=2.0.0
transformers>=4.37.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.66.0
datasets>=2.14.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
regex>=2023.6.3
accelerate>=0.24.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
huggingface_hub>=0.19.0
pdfminer.six
tiktoken
